split composable into one for listing and one for configuration
