+++ src/pages/administration/school-external-tool/SchoolExternalToolConfigurator.page.unit.ts
@@ -2,8 +2,6 @@ import SchoolExternalToolConfigurator from "./SchoolExternalToolConfigurator.pag
 import ExternalToolConfigurator from "@/components/administration/external-tools-configuration/ExternalToolConfigurator.vue";
 import ExternalToolMediumDetails from "@/components/administration/external-tools-configuration/ExternalToolMediumDetails.vue";
 import { SchoolExternalToolSave } from "@/store/external-tool";
-import SchoolExternalToolsModule from "@/store/school-external-tools";
-import { SCHOOL_EXTERNAL_TOOLS_MODULE_KEY } from "@/utils/inject";
 import { createTestAppStoreWithSchool, expectNotification } from "@@/tests/test-utils";
 import {
 	businessErrorFactory,
